feat: Naver API 를 이용한 책 검색 기능 구현

+++ woodpecker-back/src/main/java/org/example/woodpeckerback/entity/Book.java
@@ -15,7 +15,7 @@ public class Book {
     private Long id;
 
     @Column(nullable = false, unique = true)
-    private Long isbn;
+    private int isbn;
 
     @Column(nullable = false)
     private String title;
@@ -27,7 +27,7 @@ public class Book {
     private String description;
 
     @Builder
-    public Book(Long isbn, String title, String author, String publisher, String description) {
+    public Book(int isbn, String title, String author, String publisher, String description) {
         this.isbn = isbn;
         this.title = title;
         this.author = author;
